Программа по вставке в двумерный массив вещественных отрицательных чисел в цвете

+++ seminars/task017/Program.cs
@@ -1,9 +1,9 @@
 ﻿System.Random numberSintezator = new System.Random();
-int number = numberSintezator.Next(10,100);
+double number = numberSintezator.Next(10,100);
 
 Console.WriteLine(number);
-int firstNumber = number/10;
-int secondNumber = number%10;
+double firstNumber = number/10;
+double secondNumber = number%10;
 
 if(firstNumber > secondNumber)
 {
